fix(admin): resolve hydration mismatch in admin page
Apply same pattern as HeroCommunity to fix hydration mismatch:
- Extract client component to AdminPageClient.tsx
- Make page.tsx a server component that passes showPartidos prop
- Replace FeatureWrapper with conditional rendering
